backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tutoring-platform"
version = "0.1.0"
description = "Piattaforma di tutoring online completa"
authors = [{name = "Tutoring Platform Team"}]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.5",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.4.0",
    "SQLAlchemy>=2.0.34",
    "alembic>=1.13.2",
    "psycopg2-binary>=2.9.9",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "celery>=5.4.0",
    "redis>=5.0.8",
    "stripe>=8.13.0",
    "boto3>=1.34.162",
    "minio>=7.2.9",
    "emails>=0.6",
    "jinja2>=3.1.4",
    "weasyprint>=62.2",
    "httpx>=0.27.0",
    "openai>=1.42.0",
    "python-multipart>=0.0.9",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
    "pytest-mock>=3.14.0",
    "ruff>=0.6.4",
    "mypy>=1.11.2",
    "black>=23.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "minio.*",
    "stripe.*",
    "weasyprint.*",
    "emails.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
